move the federation objects to their own file
